+++ persistence/docs/CHANGELOG.md
@@ -21,6 +21,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Made `PersistenceModule` struct unexported
 - Updated tests and mocks
 - Removed some cross-module dependencies
+- Added `TxIndexer` sub-package (previously in Utility Module)
+- Added `TxIndexer` to both `PersistenceModule` and `PersistenceContext`
+- Implemented `TransactionExists` and `StoreTransaction`
 
 ## [0.0.0.6] - 2022-09-30